Move library meta test execution logic to LibraryMetaTest
